Moved api module to repository folder as well
