+++ tests/LawProposer.spec.ts
@@ -1,5 +1,5 @@
 import "should"
-import API from "../src/api"
+import API from "../src/model/api"
 import repository from "../src/model/Repository"
 import { fillUpLawProposals, replaceLawProposal } from "../src/LawProposer"
 import { Game, Law } from "../src/types"
